# openapi.yaml
openapi: 3.0.3
info:
  title: MiniBlog API
  description: |
    API REST para gestión de autores y publicaciones.
    Desarrollada con Node.js, Express y PostgreSQL.
  version: 1.0.0
  contact:
    name: DevSpark

servers:
  - url: http://localhost:3000
    description: Desarrollo local
  - url: https://<tu-app>.up.railway.app
    description: Producción (Railway)

tags:
  - name: Authors
    description: Operaciones sobre autores
  - name: Posts
    description: Operaciones sobre publicaciones

paths:

  # ──────────────────────────────────────────────
  # AUTHORS
  # ──────────────────────────────────────────────
  /authors:
    get:
      tags: [Authors]
      summary: Listar todos los authors
      responses:
        '200':
          description: Lista de authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'

    post:
      tags: [Authors]
      summary: Crear un author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
      responses:
        '201':
          description: Author creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'

  /authors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [Authors]
      summary: Obtener un author por ID
      responses:
        '200':
          description: Author encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Authors]
      summary: Actualizar un author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
      responses:
        '200':
          description: Author actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Authors]
      summary: Eliminar un author
      responses:
        '204':
          description: Eliminado correctamente
        '404':
          $ref: '#/components/responses/NotFound'

  # ──────────────────────────────────────────────
  # POSTS
  # ──────────────────────────────────────────────
  /posts:
    get:
      tags: [Posts]
      summary: Listar todos los posts
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'

    post:
      tags: [Posts]
      summary: Crear un post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Post creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /posts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [Posts]
      summary: Obtener un post por ID
      responses:
        '200':
          description: Post encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Posts]
      summary: Actualizar un post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '200':
          description: Post actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Posts]
      summary: Eliminar un post
      responses:
        '204':
          description: Eliminado correctamente
        '404':
          $ref: '#/components/responses/NotFound'

  /posts/author/{authorId}:
    parameters:
      - name: authorId
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [Posts]
      summary: Posts de un author con detalle del author (JOIN)
      responses:
        '200':
          description: Lista de posts con datos del author
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostWithAuthor'
        '404':
          $ref: '#/components/responses/NotFound'

# ──────────────────────────────────────────────
# COMPONENTES
# ──────────────────────────────────────────────
components:
  schemas:
    Author:
      type: object
      properties:
        id:         { type: integer, example: 1 }
        name:       { type: string,  example: 'Ana García' }
        email:      { type: string,  example: 'ana@example.com' }
        bio:        { type: string,  example: 'Desarrolladora full-stack' }
        created_at: { type: string,  format: date-time }

    AuthorInput:
      type: object
      required: [name, email]
      properties:
        name:  { type: string, example: 'Ana García' }
        email: { type: string, example: 'ana@example.com' }
        bio:   { type: string, example: 'Desarrolladora full-stack' }

    Post:
      type: object
      properties:
        id:         { type: integer, example: 1 }
        title:      { type: string,  example: 'Introducción a Node.js' }
        content:    { type: string,  example: 'Node.js es un runtime...' }
        author_id:  { type: integer, example: 1 }
        published:  { type: boolean, example: true }
        created_at: { type: string,  format: date-time }

    PostInput:
      type: object
      required: [title, content, author_id]
      properties:
        title:     { type: string,  example: 'Mi primer post' }
        content:   { type: string,  example: 'Contenido del post...' }
        author_id: { type: integer, example: 1 }
        published: { type: boolean, example: false }

    PostWithAuthor:
      type: object
      properties:
        id:          { type: integer }
        title:       { type: string }
        content:     { type: string }
        published:   { type: boolean }
        created_at:  { type: string, format: date-time }
        author_id:   { type: integer }
        author_name:  { type: string }
        author_email: { type: string }
        author_bio:   { type: string }

    Error:
      type: object
      properties:
        error: { type: string, example: 'Mensaje de error' }

  responses:
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'